// app/db/models/master/grades.ts
import { index, pgTable, uniqueIndex } from "drizzle-orm/pg-core";

export const grades = pgTable(
  "grades",
  (t) => ({
    id: t.char({ length: 32 }).primaryKey(),
    active: t.boolean().notNull(),
    created_at: t.bigint({ mode: "number" }),
    created_by: t.varchar({ length: 255 }),
    deleted: t.boolean().notNull(),
    last_modified_by: t.varchar({ length: 255 }),
    modified_at: t.bigint({ mode: "number" }),
    discription: t.varchar({ length: 255 }),
    grade: t.varchar({ length: 255 }),
    sort_order: t.integer().unique(),
  }),
  (table) => [index("gradesid_index").on(table.id)],
);