// app/db/models/master/subjects.ts
import { index, pgTable } from "drizzle-orm/pg-core";

export const subjects = pgTable(
  "subjects",
  (t) => ({
    id: t.char({ length: 32 }).primaryKey(),
    active: t.boolean().notNull(),
    created_at: t.bigint({ mode: "number" }),
    created_by: t.varchar({ length: 255 }),
    deleted: t.boolean().notNull(),
    last_modified_by: t.varchar({ length: 255 }),
    modified_at: t.bigint({ mode: "number" }),
    hide_sub_topics: t.boolean().default(false),
    subject: t.varchar({ length: 255 }),
    skilled_subject: t.boolean().default(true),
  }),
  (table) => [index("subjectsid_index").on(table.id)],
);